requirements.txt
requests
python-dotenv
jinja2